client interactions functionalities

+++ src/main/java/com/example/project1/SweaterWarehouse.java
@@ -1,8 +1,9 @@
 package com.example.project1;
 
+import com.example.project1.SweaterFactory.Sweater;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel = "sweater", path = "sweater")
-public interface SweaterWarehouse extends PagingAndSortingRepository<Sweater, Long> {
+public interface SweaterWarehouse extends PagingAndSortingRepository<Sweater, Integer> {
 }
